Move client definition into deepstream/client.hpp
